project_C_poisson/requirements.txt
numpy>=1.26.0
scipy>=1.10.0
pandas>=2.0.0
pyarrow>=12.0.0
scikit-learn>=1.3.0
joblib>=1.3.0
pyyaml>=6.0
